Now catching errors in try/catch
Previously anytime someone entered invalid input the errors were not being
handled at all. However TrackJS was logging any of these errors which results
in a lot of false positive errors since from the user perspective the error is
actually handled as an error message is displayed. I have added a try/catch to
ensure any error that happens the user is advised and it is caught so it is not
logged.